Extract `InputEndpointLabel.svelte` from `InputEndpoint.svelte`
